Work around bad make distfile Makefile rule
The distfile rule relies on autoconf detecting features in order to
decide what source files to include in the tarball. Since the packages
might not be available (and aren't really needed at this point of the
build), this patch adds some environment variables to force the
configure system to think all optional packages are installed and thus
include the entire set of sources in the tarball.